Make int the default Index_type

+++ include/RAJA/int_datatypes.hxx
@@ -79,7 +79,7 @@ enum IndexOwnership { Unowned, Owned };
 ///
 /// Type use for all loop indexing in RAJA constructs.
 ///
-typedef long Index_type;
+typedef int Index_type;
 
 ///
 /// Integer value for undefined indices and other integer values.
